+++ web/src/ui/Components/AppBar.tsx
@@ -23,17 +23,7 @@ import ProfileIcon from '@mui/icons-material/Person';
 import BackIcon from '@mui/icons-material/ArrowBackIosNewOutlined';
 import CircularProgress from "@mui/material/CircularProgress";
 import {shallow} from "zustand/shallow";
-import Switch from "@mui/material/Switch";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import {STAGE} from '../../utils/config'
-import dayjs from "dayjs";
-import {CREDIT_MINUTES} from '../../../../schemas/users'
-import Chip from "@mui/material/Chip";
-import {Stack2} from "./Misc.tsx";
-import Alert from "@mui/material/Alert";
-import AlertTitle from "@mui/material/AlertTitle";
-import {BasicDialog} from "./Dialog.tsx";
-import CardContent from "@mui/material/CardContent";
+import {CreditBanner} from "./CreditBanner.tsx";
 
 const buttonSx = {
   fontWeight: 300,
@@ -291,57 +281,3 @@ export default function ResponsiveAppBar({
   </>
 }
 
-function CreditBanner() {
-  const [
-    creditActive,
-    creditSeconds,
-    me,
-    setPremium
-  ] = useStore(s => [
-    s.creditActive,
-    s.creditSeconds,
-    s.user?.me,
-    s.modals.setPremium
-  ], shallow)
-  const [showShare, setShowShare] = useState(false)
-  const dialog = useMemo(() => <BasicDialog open={showShare} onClose={() => setShowShare(false)} size="sm">
-    <CardContent>
-      <Typography>Share Gnothi on social media, Reddit, etc. Email proof (a link or screenshot) to gnothi@gnothiai.com and we'll send you 7 free credits. One use per customer.</Typography>
-    </CardContent>
-  </BasicDialog>, [showShare])
-
-  if (!creditActive) { return null }
-  if (!me) { return null }
-  const timeLeft = CREDIT_MINUTES * 60 - creditSeconds
-
-  return <>
-    <Alert
-      icon={false}
-      severity="info"
-      className='credit-banner'
-      action={<Stack direction='column' gap={1}>
-        <Button variant='outlined' color="inherit" size="small" onClick={() => setPremium(true)}>
-          Upgrade for Unlimited
-        </Button>
-        <Button variant='text' color="inherit" size="small" onClick={() => setShowShare(true)}>
-          Share for +7 Credits
-        </Button>
-      </Stack>}
-      // sx={{display: "flex", justifyContent: "space-between", alignItems: "center", flexDirection: "column"}}
-    >
-      <AlertTitle className='n-credits'>{me?.credits} / 10 credits</AlertTitle>
-      <Stack2 direction='row'>
-        <Typography>Credit active:</Typography>
-        <Chip className='timer' variant="outlined" label={formatSecondsToMinutes(timeLeft)} />
-      </Stack2>
-    </Alert>
-
-    {dialog}
-  </>
-}
-
-function formatSecondsToMinutes(seconds: number): string {
-  const minutes = Math.floor(seconds / 60);
-  const remainingSeconds = seconds % 60;
-  return `${minutes}:${remainingSeconds < 10 ? '0' : ''}${remainingSeconds}`;
-}
